Split Implementation from Header Files
Properly split the implementation of functions from the header files. Changed the makefile to use a variable for the list of screens to cut down on changes that need to be made in the event of a screen being added or deleted.
