+++ src/global.c
@@ -1,0 +1,5 @@
+#include <stdlib.h>
+
+int randomInRange(int min, int max) {
+    return rand() % (max - min + 1) + min;
+}
