+++ src/global.h
@@ -1,6 +1,5 @@
 #ifndef GLOBAL_H
 #define GLOBAL_H
-#include "raylib.h"
 
 enum Screen {
     GAME,
@@ -8,8 +7,6 @@ enum Screen {
     CLOSE_GAME
 };
 
-int randomInRange(int min, int max) {
-    return rand() % (max - min + 1) + min;
-}
+int randomInRange(int min, int max);
 
 #endif
