+++ src/plinko.c
@@ -2,6 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include "raylib.h"
 #include "global.h"
 #include "screens/game.h"
 #include "screens/options.h"
